refactor(#9): 인증 관련 설정 파일 이관
추후 인증 서버 분리를 목표로 인증 관련 로직을 porko-auth 모듈로 추출하기 위해 porko-common에 위치한 인증 관련 파일을 porko-service로 이관

Controller Layer TC 수행 시 인증 정보 설정을 위한 Header 설정 부 추가
