+++ porko-service/src/test/java/io/porko/config/security/TestSecurityConfig.java
@@ -1,8 +1,8 @@
 package io.porko.config.security;
 
+import io.porko.auth.config.SecurityConfig;
 import org.springframework.context.annotation.Import;
 
-// TODO: Mocking을 통해 인증된 사용자 정보 제공
 @Import(SecurityConfig.class)
 public class TestSecurityConfig {
 }
